*Add authentication to ionic project*
The normal oauth flow won't work for ionic because on redirect the user will have no way to get back to the app, and then on auth complete the redirect can't take them back.  So instead the inAppBrowser plugin is used to open the oauth login page, and our app detects when the url changes in the in app browser and pulls the token out of it and closes the browser on auth complete.

This also requires some trickery because the process is kicked off as a POST, which browsers cannot open.  So the flow works like this:

* The in app browser opens with a data url that contains a hidden form that kicks off the oauth process (hidden fields with the client_id, response_type, etc), as well as javascript that automatically submits the form as soon as the form loads.
* Since the form is hidden and submits instantly, you'll never know this is here, but it gets us around the fact that the browser can't open a POST, so we do the post instantly within the browser.
* At this point the normal username and password and allow data access screens appear in the in app browser and the user completes them.
* When that process is complete, we detect the token being sent in the url, and as soon as loading start pull the token and expiration off the url and store them locally.
* We close the browser before loading is complete, and redirect the app to '/' home, so the main app module can load it's home page